Align the generated CMake build with the main one.
Having two of these is tedious and I hope, eventually, we can align
them. But for now, sync them manually:

- Bump the minimum CMake versions to match

- Align the C/C++ version directives

- Simplify architecture detection

- Trim some Windows defines that date to our overly aggressive warnings

- Use the Threads package

- Only use _XOPEN_SOURCE on Linux because it's a glibc-specific problem.

I've tested this manually, but we don't particularly test this build
right now (I forget if anyone is even using it), so this is mostly
relying on finding out from others if it breaks something. In the long
term, we should merge the two CMake builds.

Bug: 542
Change-Id: Icccc466464306967275d29a6982c0e9859fc972c