Clean up test_support_lib and GTest dependencies slightly.
test_support_lib depends on GTest, though only in headers. So bssl_shim
and handshaker should not depend on it. They only do for the sake of the
malloc tests. For now, split malloc.cc into its own target.

With all that cleaned up, we can remove one of the global
include_directories calls and rely on CMake's
INTERFACE_INCLUDE_DIRECTORIES machinery. It also means we can finally
depend on GTest in test_support_lib .cc files, which should be a little
less inconvenient.

Change-Id: I364df834d62328b69f146fbe35c10af97618a713