Get rid of time_t usage internally, change to int64_t
We still keep time_t stuff around for calling time() and
for external interfaces that are meant to give you time_t
values.

Update-Note: This is a potentially breaking change for things that bind to the ASN1_TIME_set and ASN1_TIME_adj family of functions (and can not type convert a time_t to an int64) Currently the only caller with an issue is easily fixed (See cl/481746965 for the fix) and would need to be fixed atomically with a google3 update.

Bug: 416

Change-Id: Ic4daba5a299d8f35191853742640750a1ecc53d6