Remove if'd-out OCB-AES assembly
BoringSSL never shipped the OCB-AES assembly, but took two differnt
strategies in disabling it for x86 versus x86_64. For x86, the
implementation was deletd, but for x86_64 it was wrapped in `if(0)`.

Since we're no longer as concerned about keeping the assembly from
diverging from upstream, be consistent in how the OCB-AES functions
are removed from both by deleting them from x86_64.

Change-Id: I5233134e3e131fed56f365ed6f43f30c39dd2e33