Revert "Drop HRSS assembly."
This reverts commit 97873cd1a59b97ced00907e274afaff75edf4a57.

In some configurations, Clang (when building for debug) compiles
poly_mul_vec_aux with a huge (12KB) stack frame. It appears to be
expanding the code into a huge SSA graph and then assigning every value
in the graph to the stack. When optimising it can store the live values
in only a few hundreds bytes of stack, but since this function recurses,
the debug build can overflow the stack.

Unclear quite what to do about this, but this change solves the
immediate issue.

Change-Id: Ifa531f91f10bf04de92e8d07fbb14c668b366454