Remove proxy certificate support.
Nothing uses this, and the code is somewhat decrepit. Instead of
fixing it and continuing to maintain it as attack surfcace, we
send this off to the farm where it can run and play all day with
the other unused X.509 extensions.

Update-note: This removes the proxy certificate extension from
the recognized certificate extensions. Previously by default
a certificate with a critical proxy certificate extension would
have been rejected with "proxy certificate not allowed", but
will now be rejected with an unrecognized critical extension
error.

Bug: 568
Change-Id: I5f838d69c59517254b4fa83f6e2abe6057fa66c7