Reflect more CI options into properties
This is in preparation for adding a static Android aarch64 FIPS build.

Right now, a builder with "android" and "fips" in the name is
automatically built as a shared library. As the first step to removing
that, specify it explicitly in the existing Android FIPS builder as
"properties". Then I'll remove the corresponding name-parsing bits from
the recipe and add them here.

While I'm here, transcribe the other CMake flags to properties. I think
I've got most of them now.

Change-Id: I08b71a8b91134ad8bb189d5a103eb6db65129ac8