Replace almost all instances of CMAKE_SOURCE_DIR and CMAKE_BINARY_DIR
If we're included as a subproject in another build, as gRPC does,
CMAKE_SOURCE_DIR points to the top-level source directory, not ours.
PROJECT_SOURCE_DIR points to ours. Likewise, CMAKE_BINARY_DIR will
point to the top-level one.

gRPC doesn't consume this CMake builds, but in preparation for
eventually unifying the two CMake builds, replacae CMAKE_SOURCE_DIR and
CMAKE_BINARY_DIR with a combination of CMAKE_CURRENT_{SOURCE,BINARY}_DIR
and PROJECT_SOURCE_DIR.

There's one more CMAKE_SOURCE_DIR which controls some default install
directory. I've left that one alone for now as I'm not sure what to do
with it. Probably the answer is to, like in gRPC, disable the install
target by default when we're not the top-level source directory.

Bug: 542
Change-Id: Iea26bbef8a4637671fd0e7476101512e871e7e18