Tidy up HMAC_Init_ex slightly
Guarding the OPENSSL_memset was unnecessary since OPENSSL_memset with
zero length works fine. Also OPENSSL_memset, to workaround a C bug,
internally does the same check anyway. (But also this wasn't a context
where the C bug applied.)

Also don't bother calling EVP_MD_block_size a second time when we
already saved it into block_size.

Finally, don't bother filling key_block up to the whole
EVP_MAX_MD_BLOCK_SIZE size. I could believe the fixed size is easier for
the compiler to optimize, but the XORs in setting up an HMAC cannot
possibly be performance-sensitive, and using the actual length is
clearer.

Also add an assert that the hash's output size fits in the block size.
We're implicitly relying on this when hashing the key down.

Change-Id: I6ce37d41ea5bdbc8890bde7910e1b5651bc35709