Give X509 an ASN1_ITEM again
I mistakenly thought no one needed X509 as an ASN1_ITEM, but that wasn't
true. wpa_supplicant relies on this. Restore this and add a test for it.

As with the rest of the rewrite, this is currently a little tedious. I'm
hoping that, as the internals are rewritten with CBS and CBB, we can
establish some cleaner patterns and abstractions.

Bug: 547
Change-Id: I761ee058f8ec916b2ec7f4730a764d46d72f1f10