Make an OPENSSL_DEPRECATED - mark OBJ_create() with it.
We may not actually want to mark OBJ_create() but I did
this as a test. Currently this works and correctly
warns for the deprecation on gcc, clang, and msvc
respectively. I have it deliberately neutered if the
gtest stuff is included so that tests may test things
marked deprecated without worrying about warnings.

Other consumers of a function we mark this way can
compile it either without -Werror, or by using
-Wno-deprecate or equivalent.

Bug: 584
Change-Id: Ia9ff386f0d22588e8a5999eda1a48b8c28dca2de