Make an OPENSSL_DEPRECATED - mark OBJ_create() with it.
We also add an OPENSSL_BEGIN_ALLOW_DEPRECATED and an
OPENSSL_END_ALLOW_DEPRECATED for testing use to
test deprecated functions.

We may not actually want to mark OBJ_create() but I did
this as a test. Currently this works and correctly
warns for the deprecation on gcc, clang, and msvc
respectively.

The purpose of this is to allow us to mark things
people should not be using as deprecated, and force some
inconvenience on the user of such things to notice them
(as opposed to a only a warning to not use it that they
may not see or read without something tripping them up.)

Bug: 584
Change-Id: Ia9ff386f0d22588e8a5999eda1a48b8c28dca2de