Add mechanism for deprecated declarations.
This allows us to mark functions as deprecated
declarations with OPENSSL_DEPRECATED.

We also add an OPENSSL_BEGIN_ALLOW_DEPRECATED and an
OPENSSL_END_ALLOW_DEPRECATED for testing use to
test deprecated functions.

The purpose of this is to allow us to mark things
people should not be using as deprecated, and force some
inconvenience on the user of such things to notice them
(as opposed to a only a warning to not use it that they
may not see or read without something tripping them up.)

The intent is to still allow use, with some effort,
before removing the function, or moving it to
libdecrepit.

We initially mark X509V3_EXT_add and X509V3_EXT_add_alias
as deprecated.

Bug: 584
Change-Id: Ia9ff386f0d22588e8a5999eda1a48b8c28dca2de