Convert X509_NAME_get_text_by_[NID|OBJ] to return UTF-8
Callers to these functions are usually using them to grab
subject name components and universally use the result as a C
string, whereas the OpenSSL versions return raw ASN1_STRING
bytes and ignore the encoding, which "usually works" in a
"hold my beer here are some bytes" sort of way until the
object is not encoded as you hoped.

Make this safer for the callers by making the returned
result be at least "text" and a C string. This converts
the ASN1_STRING bytes to UTF-8, and will introduce new
failure cases for this function if either memory allocation
fails for the UTF-8 conversion, or if the resulting UTF-8
contains a 0 codepoint and would produce an artificially
truncated C string. (truncation can still happen 'deliberately'
if too short of a buffer is passed in to hold the output).

Bug: 436
Change-Id: I487c10a5ff5188e94df520ef4c8982e593c680d7